build(llm): :heavy_plus_sign: added ollama, mypy, pydantic dependencies for LLM and static checking
Ollama package added as dependency to enable local LLM inference for user generated PCG params. Pydantic is used to enforce a schema output for the LLM and is combined with typing package. MyPy is used to do static type checking to ensure the code is documented and to minimize TypeErrors.