feat(llm): :sparkles: implemented basic LLM for generating parameters based on the user's input in schema format
Created Level dataclass to pass to LLM which contains type info for each its field. The Level dataclass contains 2 other TypedDict classes which specify the minimal attrs for layout and aesthetic, respectively. This enforces a rigid format for the information produced by the LLM while ensuring enough flexibility to tune the params appropriately. LLM is passed a user input with the level dataclass format and the LLM returns an initialized level instance that is tuned based on the inference on the input.